creación de relaciones (modelos y migraciones, todos los elementos), cambios en delete (controladores, todos menos moves), otros cambios y adaptaciones en controladores

+++ app/Controllers/Http/SubscriptionsController.ts
@@ -29,13 +29,22 @@ export default class SubscriptionsController {
         const body = request.body()
         theSubscription.subscription_start_date = body.subscription_start_date
         theSubscription.subscription_end_date = body.subscription_end_date
+        theSubscription.plan_id = body.plan_id
         return theSubscription.save()
     }
 
     public async delete({ params, response }: HttpContextContract) {
         const theSubscription: Subscription = await Subscription.findOrFail(params.id)
-        response.status(204)
-        return theSubscription.delete()
+        // response.status(204)
+        // return theSubscription.delete()
+        await theSubscription.load("payments")
+        if (theSubscription.payments) {
+            response.status(400);
+            return { "message": "Cannot be deleted because it has associated payments"}
+        } else {
+            response.status(204);
+            return theSubscription.delete();
+        }
     }
 
 }
